change meal feature list file structure
