--- src/pages/Meal/FeatureChoice/FeatureChoice.module.scss
@@ -1,50 +1,0 @@
-@import '../../../styles/variables.scss';
-
-.choice {
-    padding-bottom: 9px;
-    border-bottom: 1px solid $line;
-
-    &_blur {
-        filter: opacity(20%);
-    }
-    &__container {
-        position: relative;
-        display: flex;
-        gap: 10px;
-    }
-    &__image {
-        width: 74px;
-        height: 75px;
-        background-repeat: no-repeat;
-        background-size: cover;
-        background-position: center;
-        border-radius: 10px;
-    }
-    &__description {
-        position: relative;
-        box-sizing: content-box;
-        display: flex;
-        flex-direction: column;
-        row-gap: 10px;
-        padding: 10px 0 10px 0;
-    }
-    &__name {
-        font-size: 15px;
-        font-weight: 400;
-        line-height: 1.4;
-        color: $black;
-    }
-    &__price {
-        font-size: 17px;
-        font-weight: 500;
-        line-height: 1.4;
-        color: $light-gray;
-    }
-    &__counter {
-        position: absolute;
-        right: 0px;
-        bottom: 0px;
-        width: 110px;
-        height: 40px;
-    }
-}
