--- src/pages/Meal/FeatureChoiceLIst/FeatureChoiceList.module.scss
@@ -1,11 +1,0 @@
-@import '../../../styles/variables.scss';
-
-.choice_list {
-    overflow-y: scroll;
-    display: flex;
-    flex-direction: column;
-    row-gap: 15px;
-    border: none;
-    margin: 0;
-    padding: 0;
-}
