--- src/pages/Meal/MealFeature/MealFeature.module.scss
@@ -1,16 +1,0 @@
-@import '../../../styles/variables.scss';
-
-.feature {
-    width: 100%;
-    box-sizing: content-box;
-    display: flex;
-    flex-direction: column;
-    row-gap: 15px;
-
-    &__name {
-        font-size: 17px;
-        font-weight: 500;
-        line-height: 1.4;
-        color: $light-gray;
-    }
-}
